Move detect_manual_memcpy to its module; split up utils structs
