+++ clippy_lints/src/loops/mod.rs
@@ -6,6 +6,7 @@ mod for_mut_range_bound;
 mod for_single_element_loop;
 mod infinite_loop;
 mod manual_flatten;
+mod manual_memcpy;
 mod needless_collect;
 mod same_item_push;
 mod utils;
@@ -13,29 +14,26 @@ mod utils;
 use crate::utils::sugg::Sugg;
 use crate::utils::usage::mutated_variables;
 use crate::utils::{
-    get_enclosing_block, get_parent_expr, get_trait_def_id, higher, implements_trait, is_in_panic_handler,
-    is_integer_const, is_no_std_crate, is_refutable, is_type_diagnostic_item, last_path_segment, match_trait_method,
-    path_to_local, path_to_local_id, paths, snippet, snippet_with_applicability, span_lint, span_lint_and_help,
-    span_lint_and_sugg, sugg,
+    get_enclosing_block, get_trait_def_id, higher, implements_trait, is_in_panic_handler, is_no_std_crate,
+    is_refutable, last_path_segment, match_trait_method, path_to_local, path_to_local_id, paths,
+    snippet_with_applicability, span_lint, span_lint_and_help, span_lint_and_sugg, sugg,
 };
 use if_chain::if_chain;
-use rustc_ast::ast;
-use rustc_data_structures::fx::FxHashMap;
 use rustc_errors::Applicability;
-use rustc_hir::intravisit::{walk_block, walk_expr, walk_pat, walk_stmt, NestedVisitorMap, Visitor};
+use rustc_hir::intravisit::{walk_block, walk_expr, NestedVisitorMap, Visitor};
 use rustc_hir::{
-    BinOpKind, Block, BorrowKind, Expr, ExprKind, HirId, InlineAsmOperand, LoopSource, MatchSource, Mutability, Node,
-    Pat, PatKind, Stmt, StmtKind,
+    Block, Expr, ExprKind, HirId, InlineAsmOperand, LoopSource, MatchSource, Node, Pat, PatKind, Stmt, StmtKind,
 };
 use rustc_lint::{LateContext, LateLintPass, LintContext};
 use rustc_middle::hir::map::Map;
 use rustc_middle::lint::in_external_macro;
-use rustc_middle::ty::{self, Ty};
 use rustc_session::{declare_lint_pass, declare_tool_lint};
 use rustc_span::source_map::Span;
-use rustc_span::symbol::{sym, Symbol};
+use rustc_span::symbol::sym;
 use std::iter::{once, Iterator};
-use utils::{get_span_of_entire_for_loop, make_iterator_snippet};
+use utils::{
+    get_span_of_entire_for_loop, make_iterator_snippet, IncrementVisitor, InitializeVisitor, LoopNestVisitor, Nesting,
+};
 
 declare_clippy_lint! {
     /// **What it does:** Checks for for-loops that manually copy items between
@@ -857,7 +855,7 @@ fn check_for_loop<'tcx>(
     expr: &'tcx Expr<'_>,
     span: Span,
 ) {
-    let is_manual_memcpy_triggered = detect_manual_memcpy(cx, pat, arg, body, expr);
+    let is_manual_memcpy_triggered = manual_memcpy::detect_manual_memcpy(cx, pat, arg, body, expr);
     if !is_manual_memcpy_triggered {
         for_loop_range::check_for_loop_range(cx, pat, arg, body, expr);
         for_loop_explicit_counter::check_for_loop_explicit_counter(cx, pat, arg, body, expr);
@@ -938,373 +936,6 @@ impl std::ops::Sub<&MinifyingSugg<'static>> for MinifyingSugg<'static> {
             (_, _) => (self.0 - &rhs.0).into(),
         }
     }
-}
-
-/// a wrapper around `MinifyingSugg`, which carries a operator like currying
-/// so that the suggested code become more efficient (e.g. `foo + -bar` `foo - bar`).
-struct Offset {
-    value: MinifyingSugg<'static>,
-    sign: OffsetSign,
-}
-
-#[derive(Clone, Copy)]
-enum OffsetSign {
-    Positive,
-    Negative,
-}
-
-impl Offset {
-    fn negative(value: Sugg<'static>) -> Self {
-        Self {
-            value: value.into(),
-            sign: OffsetSign::Negative,
-        }
-    }
-
-    fn positive(value: Sugg<'static>) -> Self {
-        Self {
-            value: value.into(),
-            sign: OffsetSign::Positive,
-        }
-    }
-
-    fn empty() -> Self {
-        Self::positive(sugg::ZERO)
-    }
-}
-
-fn apply_offset(lhs: &MinifyingSugg<'static>, rhs: &Offset) -> MinifyingSugg<'static> {
-    match rhs.sign {
-        OffsetSign::Positive => lhs + &rhs.value,
-        OffsetSign::Negative => lhs - &rhs.value,
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum StartKind<'hir> {
-    Range,
-    Counter { initializer: &'hir Expr<'hir> },
-}
-
-struct IndexExpr<'hir> {
-    base: &'hir Expr<'hir>,
-    idx: StartKind<'hir>,
-    idx_offset: Offset,
-}
-
-struct Start<'hir> {
-    id: HirId,
-    kind: StartKind<'hir>,
-}
-
-fn is_slice_like<'tcx>(cx: &LateContext<'tcx>, ty: Ty<'_>) -> bool {
-    let is_slice = match ty.kind() {
-        ty::Ref(_, subty, _) => is_slice_like(cx, subty),
-        ty::Slice(..) | ty::Array(..) => true,
-        _ => false,
-    };
-
-    is_slice || is_type_diagnostic_item(cx, ty, sym::vec_type) || is_type_diagnostic_item(cx, ty, sym!(vecdeque_type))
-}
-
-fn fetch_cloned_expr<'tcx>(expr: &'tcx Expr<'tcx>) -> &'tcx Expr<'tcx> {
-    if_chain! {
-        if let ExprKind::MethodCall(method, _, args, _) = expr.kind;
-        if method.ident.name == sym::clone;
-        if args.len() == 1;
-        if let Some(arg) = args.get(0);
-        then { arg } else { expr }
-    }
-}
-
-fn get_details_from_idx<'tcx>(
-    cx: &LateContext<'tcx>,
-    idx: &Expr<'_>,
-    starts: &[Start<'tcx>],
-) -> Option<(StartKind<'tcx>, Offset)> {
-    fn get_start<'tcx>(e: &Expr<'_>, starts: &[Start<'tcx>]) -> Option<StartKind<'tcx>> {
-        let id = path_to_local(e)?;
-        starts.iter().find(|start| start.id == id).map(|start| start.kind)
-    }
-
-    fn get_offset<'tcx>(cx: &LateContext<'tcx>, e: &Expr<'_>, starts: &[Start<'tcx>]) -> Option<Sugg<'static>> {
-        match &e.kind {
-            ExprKind::Lit(l) => match l.node {
-                ast::LitKind::Int(x, _ty) => Some(Sugg::NonParen(x.to_string().into())),
-                _ => None,
-            },
-            ExprKind::Path(..) if get_start(e, starts).is_none() => Some(Sugg::hir(cx, e, "???")),
-            _ => None,
-        }
-    }
-
-    match idx.kind {
-        ExprKind::Binary(op, lhs, rhs) => match op.node {
-            BinOpKind::Add => {
-                let offset_opt = get_start(lhs, starts)
-                    .and_then(|s| get_offset(cx, rhs, starts).map(|o| (s, o)))
-                    .or_else(|| get_start(rhs, starts).and_then(|s| get_offset(cx, lhs, starts).map(|o| (s, o))));
-
-                offset_opt.map(|(s, o)| (s, Offset::positive(o)))
-            },
-            BinOpKind::Sub => {
-                get_start(lhs, starts).and_then(|s| get_offset(cx, rhs, starts).map(|o| (s, Offset::negative(o))))
-            },
-            _ => None,
-        },
-        ExprKind::Path(..) => get_start(idx, starts).map(|s| (s, Offset::empty())),
-        _ => None,
-    }
-}
-
-fn get_assignment<'tcx>(e: &'tcx Expr<'tcx>) -> Option<(&'tcx Expr<'tcx>, &'tcx Expr<'tcx>)> {
-    if let ExprKind::Assign(lhs, rhs, _) = e.kind {
-        Some((lhs, rhs))
-    } else {
-        None
-    }
-}
-
-/// Get assignments from the given block.
-/// The returned iterator yields `None` if no assignment expressions are there,
-/// filtering out the increments of the given whitelisted loop counters;
-/// because its job is to make sure there's nothing other than assignments and the increments.
-fn get_assignments<'a, 'tcx>(
-    Block { stmts, expr, .. }: &'tcx Block<'tcx>,
-    loop_counters: &'a [Start<'tcx>],
-) -> impl Iterator<Item = Option<(&'tcx Expr<'tcx>, &'tcx Expr<'tcx>)>> + 'a {
-    // As the `filter` and `map` below do different things, I think putting together
-    // just increases complexity. (cc #3188 and #4193)
-    stmts
-        .iter()
-        .filter_map(move |stmt| match stmt.kind {
-            StmtKind::Local(..) | StmtKind::Item(..) => None,
-            StmtKind::Expr(e) | StmtKind::Semi(e) => Some(e),
-        })
-        .chain((*expr).into_iter())
-        .filter(move |e| {
-            if let ExprKind::AssignOp(_, place, _) = e.kind {
-                path_to_local(place).map_or(false, |id| {
-                    !loop_counters
-                        .iter()
-                        // skip the first item which should be `StartKind::Range`
-                        // this makes it possible to use the slice with `StartKind::Range` in the same iterator loop.
-                        .skip(1)
-                        .any(|counter| counter.id == id)
-                })
-            } else {
-                true
-            }
-        })
-        .map(get_assignment)
-}
-
-fn get_loop_counters<'a, 'tcx>(
-    cx: &'a LateContext<'tcx>,
-    body: &'tcx Block<'tcx>,
-    expr: &'tcx Expr<'_>,
-) -> Option<impl Iterator<Item = Start<'tcx>> + 'a> {
-    // Look for variables that are incremented once per loop iteration.
-    let mut increment_visitor = IncrementVisitor::new(cx);
-    walk_block(&mut increment_visitor, body);
-
-    // For each candidate, check the parent block to see if
-    // it's initialized to zero at the start of the loop.
-    get_enclosing_block(&cx, expr.hir_id).and_then(|block| {
-        increment_visitor
-            .into_results()
-            .filter_map(move |var_id| {
-                let mut initialize_visitor = InitializeVisitor::new(cx, expr, var_id);
-                walk_block(&mut initialize_visitor, block);
-
-                initialize_visitor.get_result().map(|(_, initializer)| Start {
-                    id: var_id,
-                    kind: StartKind::Counter { initializer },
-                })
-            })
-            .into()
-    })
-}
-
-fn build_manual_memcpy_suggestion<'tcx>(
-    cx: &LateContext<'tcx>,
-    start: &Expr<'_>,
-    end: &Expr<'_>,
-    limits: ast::RangeLimits,
-    dst: &IndexExpr<'_>,
-    src: &IndexExpr<'_>,
-) -> String {
-    fn print_offset(offset: MinifyingSugg<'static>) -> MinifyingSugg<'static> {
-        if offset.as_str() == "0" {
-            sugg::EMPTY.into()
-        } else {
-            offset
-        }
-    }
-
-    let print_limit = |end: &Expr<'_>, end_str: &str, base: &Expr<'_>, sugg: MinifyingSugg<'static>| {
-        if_chain! {
-            if let ExprKind::MethodCall(method, _, len_args, _) = end.kind;
-            if method.ident.name == sym!(len);
-            if len_args.len() == 1;
-            if let Some(arg) = len_args.get(0);
-            if path_to_local(arg) == path_to_local(base);
-            then {
-                if sugg.as_str() == end_str {
-                    sugg::EMPTY.into()
-                } else {
-                    sugg
-                }
-            } else {
-                match limits {
-                    ast::RangeLimits::Closed => {
-                        sugg + &sugg::ONE.into()
-                    },
-                    ast::RangeLimits::HalfOpen => sugg,
-                }
-            }
-        }
-    };
-
-    let start_str = Sugg::hir(cx, start, "").into();
-    let end_str: MinifyingSugg<'_> = Sugg::hir(cx, end, "").into();
-
-    let print_offset_and_limit = |idx_expr: &IndexExpr<'_>| match idx_expr.idx {
-        StartKind::Range => (
-            print_offset(apply_offset(&start_str, &idx_expr.idx_offset)).into_sugg(),
-            print_limit(
-                end,
-                end_str.as_str(),
-                idx_expr.base,
-                apply_offset(&end_str, &idx_expr.idx_offset),
-            )
-            .into_sugg(),
-        ),
-        StartKind::Counter { initializer } => {
-            let counter_start = Sugg::hir(cx, initializer, "").into();
-            (
-                print_offset(apply_offset(&counter_start, &idx_expr.idx_offset)).into_sugg(),
-                print_limit(
-                    end,
-                    end_str.as_str(),
-                    idx_expr.base,
-                    apply_offset(&end_str, &idx_expr.idx_offset) + &counter_start - &start_str,
-                )
-                .into_sugg(),
-            )
-        },
-    };
-
-    let (dst_offset, dst_limit) = print_offset_and_limit(&dst);
-    let (src_offset, src_limit) = print_offset_and_limit(&src);
-
-    let dst_base_str = snippet(cx, dst.base.span, "???");
-    let src_base_str = snippet(cx, src.base.span, "???");
-
-    let dst = if dst_offset == sugg::EMPTY && dst_limit == sugg::EMPTY {
-        dst_base_str
-    } else {
-        format!(
-            "{}[{}..{}]",
-            dst_base_str,
-            dst_offset.maybe_par(),
-            dst_limit.maybe_par()
-        )
-        .into()
-    };
-
-    format!(
-        "{}.clone_from_slice(&{}[{}..{}]);",
-        dst,
-        src_base_str,
-        src_offset.maybe_par(),
-        src_limit.maybe_par()
-    )
-}
-
-/// Checks for for loops that sequentially copy items from one slice-like
-/// object to another.
-fn detect_manual_memcpy<'tcx>(
-    cx: &LateContext<'tcx>,
-    pat: &'tcx Pat<'_>,
-    arg: &'tcx Expr<'_>,
-    body: &'tcx Expr<'_>,
-    expr: &'tcx Expr<'_>,
-) -> bool {
-    if let Some(higher::Range {
-        start: Some(start),
-        end: Some(end),
-        limits,
-    }) = higher::range(arg)
-    {
-        // the var must be a single name
-        if let PatKind::Binding(_, canonical_id, _, _) = pat.kind {
-            let mut starts = vec![Start {
-                id: canonical_id,
-                kind: StartKind::Range,
-            }];
-
-            // This is one of few ways to return different iterators
-            // derived from: https://stackoverflow.com/questions/29760668/conditionally-iterate-over-one-of-several-possible-iterators/52064434#52064434
-            let mut iter_a = None;
-            let mut iter_b = None;
-
-            if let ExprKind::Block(block, _) = body.kind {
-                if let Some(loop_counters) = get_loop_counters(cx, block, expr) {
-                    starts.extend(loop_counters);
-                }
-                iter_a = Some(get_assignments(block, &starts));
-            } else {
-                iter_b = Some(get_assignment(body));
-            }
-
-            let assignments = iter_a.into_iter().flatten().chain(iter_b.into_iter());
-
-            let big_sugg = assignments
-                // The only statements in the for loops can be indexed assignments from
-                // indexed retrievals (except increments of loop counters).
-                .map(|o| {
-                    o.and_then(|(lhs, rhs)| {
-                        let rhs = fetch_cloned_expr(rhs);
-                        if_chain! {
-                            if let ExprKind::Index(base_left, idx_left) = lhs.kind;
-                            if let ExprKind::Index(base_right, idx_right) = rhs.kind;
-                            if is_slice_like(cx, cx.typeck_results().expr_ty(base_left))
-                                && is_slice_like(cx, cx.typeck_results().expr_ty(base_right));
-                            if let Some((start_left, offset_left)) = get_details_from_idx(cx, &idx_left, &starts);
-                            if let Some((start_right, offset_right)) = get_details_from_idx(cx, &idx_right, &starts);
-
-                            // Source and destination must be different
-                            if path_to_local(base_left) != path_to_local(base_right);
-                            then {
-                                Some((IndexExpr { base: base_left, idx: start_left, idx_offset: offset_left },
-                                    IndexExpr { base: base_right, idx: start_right, idx_offset: offset_right }))
-                            } else {
-                                None
-                            }
-                        }
-                    })
-                })
-                .map(|o| o.map(|(dst, src)| build_manual_memcpy_suggestion(cx, start, end, limits, &dst, &src)))
-                .collect::<Option<Vec<_>>>()
-                .filter(|v| !v.is_empty())
-                .map(|v| v.join("\n    "));
-
-            if let Some(big_sugg) = big_sugg {
-                span_lint_and_sugg(
-                    cx,
-                    MANUAL_MEMCPY,
-                    get_span_of_entire_for_loop(expr),
-                    "it looks like you're manually copying between slices",
-                    "try replacing the loop by",
-                    big_sugg,
-                    Applicability::Unspecified,
-                );
-                return true;
-            }
-        }
-    }
-    false
 }
 
 fn is_used_inside<'tcx>(cx: &LateContext<'tcx>, expr: &'tcx Expr<'_>, container: &'tcx Expr<'_>) -> bool {
@@ -1398,233 +1029,6 @@ fn is_simple_break_expr(expr: &Expr<'_>) -> bool {
     }
 }
 
-#[derive(Debug, PartialEq)]
-enum IncrementVisitorVarState {
-    Initial,  // Not examined yet
-    IncrOnce, // Incremented exactly once, may be a loop counter
-    DontWarn,
-}
-
-/// Scan a for loop for variables that are incremented exactly once and not used after that.
-struct IncrementVisitor<'a, 'tcx> {
-    cx: &'a LateContext<'tcx>,                          // context reference
-    states: FxHashMap<HirId, IncrementVisitorVarState>, // incremented variables
-    depth: u32,                                         // depth of conditional expressions
-    done: bool,
-}
-
-impl<'a, 'tcx> IncrementVisitor<'a, 'tcx> {
-    fn new(cx: &'a LateContext<'tcx>) -> Self {
-        Self {
-            cx,
-            states: FxHashMap::default(),
-            depth: 0,
-            done: false,
-        }
-    }
-
-    fn into_results(self) -> impl Iterator<Item = HirId> {
-        self.states.into_iter().filter_map(|(id, state)| {
-            if state == IncrementVisitorVarState::IncrOnce {
-                Some(id)
-            } else {
-                None
-            }
-        })
-    }
-}
-
-impl<'a, 'tcx> Visitor<'tcx> for IncrementVisitor<'a, 'tcx> {
-    type Map = Map<'tcx>;
-
-    fn visit_expr(&mut self, expr: &'tcx Expr<'_>) {
-        if self.done {
-            return;
-        }
-
-        // If node is a variable
-        if let Some(def_id) = path_to_local(expr) {
-            if let Some(parent) = get_parent_expr(self.cx, expr) {
-                let state = self.states.entry(def_id).or_insert(IncrementVisitorVarState::Initial);
-                if *state == IncrementVisitorVarState::IncrOnce {
-                    *state = IncrementVisitorVarState::DontWarn;
-                    return;
-                }
-
-                match parent.kind {
-                    ExprKind::AssignOp(op, ref lhs, ref rhs) => {
-                        if lhs.hir_id == expr.hir_id {
-                            *state = if op.node == BinOpKind::Add
-                                && is_integer_const(self.cx, rhs, 1)
-                                && *state == IncrementVisitorVarState::Initial
-                                && self.depth == 0
-                            {
-                                IncrementVisitorVarState::IncrOnce
-                            } else {
-                                // Assigned some other value or assigned multiple times
-                                IncrementVisitorVarState::DontWarn
-                            };
-                        }
-                    },
-                    ExprKind::Assign(ref lhs, _, _) if lhs.hir_id == expr.hir_id => {
-                        *state = IncrementVisitorVarState::DontWarn
-                    },
-                    ExprKind::AddrOf(BorrowKind::Ref, mutability, _) if mutability == Mutability::Mut => {
-                        *state = IncrementVisitorVarState::DontWarn
-                    },
-                    _ => (),
-                }
-            }
-
-            walk_expr(self, expr);
-        } else if is_loop(expr) || is_conditional(expr) {
-            self.depth += 1;
-            walk_expr(self, expr);
-            self.depth -= 1;
-        } else if let ExprKind::Continue(_) = expr.kind {
-            self.done = true;
-        } else {
-            walk_expr(self, expr);
-        }
-    }
-    fn nested_visit_map(&mut self) -> NestedVisitorMap<Self::Map> {
-        NestedVisitorMap::None
-    }
-}
-
-enum InitializeVisitorState<'hir> {
-    Initial,          // Not examined yet
-    Declared(Symbol), // Declared but not (yet) initialized
-    Initialized {
-        name: Symbol,
-        initializer: &'hir Expr<'hir>,
-    },
-    DontWarn,
-}
-
-/// Checks whether a variable is initialized at the start of a loop and not modified
-/// and used after the loop.
-struct InitializeVisitor<'a, 'tcx> {
-    cx: &'a LateContext<'tcx>,  // context reference
-    end_expr: &'tcx Expr<'tcx>, // the for loop. Stop scanning here.
-    var_id: HirId,
-    state: InitializeVisitorState<'tcx>,
-    depth: u32, // depth of conditional expressions
-    past_loop: bool,
-}
-
-impl<'a, 'tcx> InitializeVisitor<'a, 'tcx> {
-    fn new(cx: &'a LateContext<'tcx>, end_expr: &'tcx Expr<'tcx>, var_id: HirId) -> Self {
-        Self {
-            cx,
-            end_expr,
-            var_id,
-            state: InitializeVisitorState::Initial,
-            depth: 0,
-            past_loop: false,
-        }
-    }
-
-    fn get_result(&self) -> Option<(Symbol, &'tcx Expr<'tcx>)> {
-        if let InitializeVisitorState::Initialized { name, initializer } = self.state {
-            Some((name, initializer))
-        } else {
-            None
-        }
-    }
-}
-
-impl<'a, 'tcx> Visitor<'tcx> for InitializeVisitor<'a, 'tcx> {
-    type Map = Map<'tcx>;
-
-    fn visit_stmt(&mut self, stmt: &'tcx Stmt<'_>) {
-        // Look for declarations of the variable
-        if_chain! {
-            if let StmtKind::Local(ref local) = stmt.kind;
-            if local.pat.hir_id == self.var_id;
-            if let PatKind::Binding(.., ident, _) = local.pat.kind;
-            then {
-                self.state = local.init.map_or(InitializeVisitorState::Declared(ident.name), |init| {
-                    InitializeVisitorState::Initialized {
-                        initializer: init,
-                        name: ident.name,
-                    }
-                })
-            }
-        }
-        walk_stmt(self, stmt);
-    }
-
-    fn visit_expr(&mut self, expr: &'tcx Expr<'_>) {
-        if matches!(self.state, InitializeVisitorState::DontWarn) {
-            return;
-        }
-        if expr.hir_id == self.end_expr.hir_id {
-            self.past_loop = true;
-            return;
-        }
-        // No need to visit expressions before the variable is
-        // declared
-        if matches!(self.state, InitializeVisitorState::Initial) {
-            return;
-        }
-
-        // If node is the desired variable, see how it's used
-        if path_to_local_id(expr, self.var_id) {
-            if self.past_loop {
-                self.state = InitializeVisitorState::DontWarn;
-                return;
-            }
-
-            if let Some(parent) = get_parent_expr(self.cx, expr) {
-                match parent.kind {
-                    ExprKind::AssignOp(_, ref lhs, _) if lhs.hir_id == expr.hir_id => {
-                        self.state = InitializeVisitorState::DontWarn;
-                    },
-                    ExprKind::Assign(ref lhs, ref rhs, _) if lhs.hir_id == expr.hir_id => {
-                        self.state = if_chain! {
-                            if self.depth == 0;
-                            if let InitializeVisitorState::Declared(name)
-                                | InitializeVisitorState::Initialized { name, ..} = self.state;
-                            then {
-                                InitializeVisitorState::Initialized { initializer: rhs, name }
-                            } else {
-                                InitializeVisitorState::DontWarn
-                            }
-                        }
-                    },
-                    ExprKind::AddrOf(BorrowKind::Ref, mutability, _) if mutability == Mutability::Mut => {
-                        self.state = InitializeVisitorState::DontWarn
-                    },
-                    _ => (),
-                }
-            }
-
-            walk_expr(self, expr);
-        } else if !self.past_loop && is_loop(expr) {
-            self.state = InitializeVisitorState::DontWarn;
-        } else if is_conditional(expr) {
-            self.depth += 1;
-            walk_expr(self, expr);
-            self.depth -= 1;
-        } else {
-            walk_expr(self, expr);
-        }
-    }
-
-    fn nested_visit_map(&mut self) -> NestedVisitorMap<Self::Map> {
-        NestedVisitorMap::OnlyBodies(self.cx.tcx.hir())
-    }
-}
-
-fn is_loop(expr: &Expr<'_>) -> bool {
-    matches!(expr.kind, ExprKind::Loop(..))
-}
-
-fn is_conditional(expr: &Expr<'_>) -> bool {
-    matches!(expr.kind, ExprKind::If(..) | ExprKind::Match(..))
-}
-
 fn is_nested(cx: &LateContext<'_>, match_expr: &Expr<'_>, iter_expr: &Expr<'_>) -> bool {
     if_chain! {
         if let Some(loop_block) = get_enclosing_block(cx, match_expr.hir_id);
@@ -1659,10 +1063,10 @@ fn is_loop_nested(cx: &LateContext<'_>, loop_expr: &Expr<'_>, iter_expr: &Expr<'
                 let mut block_visitor = LoopNestVisitor {
                     hir_id: id,
                     iterator: iter_id,
-                    nesting: Unknown,
+                    nesting: Nesting::Unknown,
                 };
                 walk_block(&mut block_visitor, block);
-                if block_visitor.nesting == RuledOut {
+                if block_visitor.nesting == Nesting::RuledOut {
                     return false;
                 }
             },
@@ -1672,67 +1076,5 @@ fn is_loop_nested(cx: &LateContext<'_>, loop_expr: &Expr<'_>, iter_expr: &Expr<'
             },
         }
         id = parent;
-    }
-}
-
-#[derive(PartialEq, Eq)]
-enum Nesting {
-    Unknown,     // no nesting detected yet
-    RuledOut,    // the iterator is initialized or assigned within scope
-    LookFurther, // no nesting detected, no further walk required
-}
-
-use self::Nesting::{LookFurther, RuledOut, Unknown};
-
-struct LoopNestVisitor {
-    hir_id: HirId,
-    iterator: HirId,
-    nesting: Nesting,
-}
-
-impl<'tcx> Visitor<'tcx> for LoopNestVisitor {
-    type Map = Map<'tcx>;
-
-    fn visit_stmt(&mut self, stmt: &'tcx Stmt<'_>) {
-        if stmt.hir_id == self.hir_id {
-            self.nesting = LookFurther;
-        } else if self.nesting == Unknown {
-            walk_stmt(self, stmt);
-        }
-    }
-
-    fn visit_expr(&mut self, expr: &'tcx Expr<'_>) {
-        if self.nesting != Unknown {
-            return;
-        }
-        if expr.hir_id == self.hir_id {
-            self.nesting = LookFurther;
-            return;
-        }
-        match expr.kind {
-            ExprKind::Assign(ref path, _, _) | ExprKind::AssignOp(_, ref path, _) => {
-                if path_to_local_id(path, self.iterator) {
-                    self.nesting = RuledOut;
-                }
-            },
-            _ => walk_expr(self, expr),
-        }
-    }
-
-    fn visit_pat(&mut self, pat: &'tcx Pat<'_>) {
-        if self.nesting != Unknown {
-            return;
-        }
-        if let PatKind::Binding(_, id, ..) = pat.kind {
-            if id == self.iterator {
-                self.nesting = RuledOut;
-                return;
-            }
-        }
-        walk_pat(self, pat)
-    }
-
-    fn nested_visit_map(&mut self) -> NestedVisitorMap<Self::Map> {
-        NestedVisitorMap::None
     }
 }
